refactor remittance percentage
