+++ server/schemas/referral.schema.js
@@ -1,10 +1,5 @@
 import Joi from '@hapi/joi';
-import {
-  requiredEmail,
-  requiredObjectId,
-  optionalString,
-  requiredPercentage,
-} from './helper.schema';
+import { requiredEmail, requiredObjectId, optionalString } from './helper.schema';
 
 export const updateReferralSchema = Joi.object({
   referralId: requiredObjectId('Referral Id'),
@@ -13,9 +8,4 @@ export const updateReferralSchema = Joi.object({
 export const sendReferralSchema = Joi.object({
   email: requiredEmail('Email Address'),
   firstName: optionalString('First Name'),
-});
-
-export const updateReferralPercentageSchema = Joi.object({
-  referralId: requiredObjectId('Referral Id'),
-  percentage: requiredPercentage('Percentage'),
 });
