+++ test/controllers/referral.controller.test.js
@@ -911,102 +911,4 @@ describe('Referral Controller', () => {
       });
     });
   });
-
-  describe('Reward a referral', () => {
-    const method = 'put';
-    const endpoint = '/api/v1/referral/percentage';
-
-    const referral = ReferralFactory.build(
-      { referrerId: adminUser._id, 'reward.status': REWARD_STATUS.PAYMENT_IN_PROGRESS },
-      { generateId: true },
-    );
-    const vendorUser = UserFactory.build(
-      { role: USER_ROLE.VENDOR, activated: true },
-      { generateId: true },
-    );
-
-    const referralDetails = {
-      referralId: referral._id,
-      percentage: 2,
-    };
-
-    beforeEach(async () => {
-      await addUser(vendorUser);
-      await addReferral(referral);
-    });
-
-    context('with valid data & token', () => {
-      it('updates referral percentage', (done) => {
-        request()
-          .put(endpoint)
-          .set('authorization', adminToken)
-          .send(referralDetails)
-          .end((err, res) => {
-            expect(res).to.have.status(200);
-            expect(res.body.success).to.be.eql(true);
-            expect(res.body.message).to.be.eql('Referral percentage updated');
-            expect(res.body.referral._id).to.be.eql(referral._id.toString());
-            expect(res.body.referral.reward.percentage).to.be.eql(referralDetails.percentage);
-            done();
-          });
-      });
-    });
-
-    context('when referral has been paid', () => {
-      beforeEach(async () => {
-        await Referral.findByIdAndUpdate(referral._id, {
-          'reward.status': REWARD_STATUS.REFERRAL_PAID,
-        });
-      });
-
-      it('returns error', (done) => {
-        request()
-          .put(endpoint)
-          .set('authorization', adminToken)
-          .send(referralDetails)
-          .end((err, res) => {
-            expect(res).to.have.status(412);
-            expect(res.body.success).to.be.eql(false);
-            expect(res.body.message).to.be.eql('Referral has been paid');
-            done();
-          });
-      });
-    });
-
-    context('with invalid referral id', () => {
-      it('returns not found', (done) => {
-        request()
-          .put(endpoint)
-          .set('authorization', adminToken)
-          .send({ ...referralDetails, referralId: mongoose.Types.ObjectId() })
-          .end((err, res) => {
-            expect(res.body.success).to.be.eql(false);
-            expect(res.body.error.statusCode).to.be.eql(404);
-            expect(res.body.error.message).to.be.eql('Referral not found');
-            done();
-          });
-      });
-    });
-
-    itReturnsForbiddenForNoToken({ endpoint, method, data: referralDetails });
-
-    [regularUser, vendorUser].map((user) =>
-      itReturnsForbiddenForTokenWithInvalidAccess({
-        endpoint,
-        method,
-        user,
-        data: referralDetails,
-        useExistingUser: true,
-      }),
-    );
-
-    itReturnsNotFoundForInvalidToken({
-      endpoint,
-      method,
-      user: adminUser,
-      userId: adminUser._id,
-      data: referralDetails,
-      useExistingUser: true,
-    });
-  });
 });
